requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
babel==2.17.0
beautifulsoup4==4.13.5
blinker==1.9.0
brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.1
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
et-xmlfile==2.0.0
fastapi==0.116.2
flask==3.1.2
fonttools==4.60.0
fpdf==1.7.2
fpdf2==2.8.4
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.1
markdown==3.9
markupsafe==3.0.2
matplotlib==3.10.6
narwhals==2.5.0
numpy==2.3.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pillow==11.3.0
playwright==1.55.0
plotly==6.3.0
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydyf==0.11.0
pyee==13.0.0
pyhanko==0.31.0
pyhanko-certvalidator==0.29.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyngrok==7.4.0
pyparsing==3.2.4
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
qrcode==8.2
reportlab==4.4.3
requests==2.32.5
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tkinterweb==4.4.4
tkinterweb-tkhtml==1.1.4
tqdm==4.67.1
ttkthemes==3.2.2
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
